Update lazy setting configuration and methods
The commit includes the update of the 'cache_key' to 'cache_prefix' in the lazys setting file. A new 'config' method has been introduced in LazySetting.php to fetch configurations. Furthermore, the logic of the 'getCacheKey' method has also been adjusted in line with these changes. Lastly, a 'getTable' method has been added to the Setting model to retrieve the table configuration.